shared/requirements.txt
pydantic-settings>=2.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aio-pika>=9.0.0
redis[asyncio]>=5.0.0
aioboto3>=12.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-exporter-jaeger>=1.20.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
